Move grid_cells_display.hpp to include folder
