+++ rviz_default_plugins/test/rviz_default_plugins/displays/grid_cells/grid_cells_display_test.cpp
@@ -38,7 +38,7 @@
 
 #include "rviz_common/properties/float_property.hpp"
 
-#include "../../../../src/rviz_default_plugins/displays/grid_cells/grid_cells_display.hpp"
+#include "rviz_default_plugins/displays/grid_cells/grid_cells_display.hpp"
 
 #include "../display_test_fixture.hpp"
 #include "../../scene_graph_introspection.hpp"
